Assistant checkpoint: Update screen_time_earned column type to FLOAT
Assistant generated file changes:
- migrations/0005_fix_screen_time_earned_type.sql: Add migration to change screen_time_earned to FLOAT
- db/schema.ts: Update schema to use float type

---

User prompt:

game_results table screen_time_earns is TEXT data type, it should be FLOAT, could you make the change w/o having to lose data?

+++ db/schema.ts
@@ -23,7 +23,7 @@ export const gameResults = pgTable('game_results', {
   timeTakenInMs: integer('time_taken_in_ms').notNull(),
   bestStreak: integer('best_streak').notNull(),
   incorrectAttempts: integer('incorrect_attempts').notNull(),
-  screenTimeEarned: text('screen_time_earned'),
+  screenTimeEarned: text('screen_time_earned').type('float'),
   createdAt: text('created_at')
     .default(sql`CURRENT_TIMESTAMP`)
     .notNull(),
